AEA-367: Corrected Postman test. Tidied up request handler utility functions.
