+++ coordinator/src/routes/prescription/post-convert-full-message.ts
@@ -1,7 +1,7 @@
-import {validatingHandler} from "../../services/handler";
 import * as translator from "../../services/translation/translation-service";
 import Hapi from "@hapi/hapi";
 import {Bundle} from "../../model/fhir-resources";
+import {validatingHandler} from "../util";
 
 const CONTENT_TYPE = 'application/xml'
